Rename module & fix or suppress mypy errors
