+++ vehicle-python-core/.python-version
@@ -1,0 +1,5 @@
+3.7.16
+3.8.16
+3.9.16
+3.10.10
+3.11.2
