+++ vehicle-python-core/Setup.hs
@@ -12,7 +12,7 @@ import Distribution.Simple.Utils (die', info)
 import Distribution.Types.LocalBuildInfo (componentNameCLBIs)
 import Distribution.Types.UnqualComponentName (UnqualComponentName, unUnqualComponentName)
 import Distribution.Verbosity (Verbosity, normal)
-import System.Directory (doesDirectoryExist, copyFile)
+import System.Directory (copyFile, doesDirectoryExist)
 import System.Environment (getEnv)
 import System.FilePath ((<.>))
 import System.FilePath.Posix ((</>))
