+++ vehicle-python-core/pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=45"]
+requires = ["setuptools >=45"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -12,9 +12,15 @@ license = { file = 'LICENSE' }
 
 [project.optional-dependencies]
 test = ["pytest >=7.1,<8"]
+mypy = ["types-setuptools >=45"]
 
 [project.scripts]
-vehicle = "vehicle.core:main"
+vehicle = "vehicle_core:main"
+
+[tool.mypy]
+python_version = 3.7
+ignore_missing_imports = true
+exclude = ['^src\/vehicle_core\/binding\.py$']
 
 [tool.tox]
 legacy_tox_ini = """
