+++ vehicle-python-core/setup.py
@@ -1,20 +1,23 @@
 import os.path
-import sys
 import subprocess
+import sys
 from typing import Optional
+
 from setuptools import Extension, setup
 from setuptools.command.build_ext import build_ext
-from distutils.errors import (
-    DistutilsExecError,
-    DistutilsPlatformError,
-    DistutilsSetupError,
-)
-from distutils.spawn import find_executable
+
+# 03-04-2023:
+# The imports from distutils must come AFTER the imports from setuptools,
+# because setuptools overrides the distutils modules with its own versions.
+# isort: split
+
+import distutils.errors
+import distutils.spawn
 
 ext_modules = [
     Extension(
-        name="vehicle.core._binding",
-        sources=["src/vehicle/core/binding.i"],
+        name="vehicle_core._binding",
+        sources=["src/vehicle_core/binding.i"],
     ),
 ]
 
@@ -29,13 +32,13 @@ class cabal_build_ext(build_ext):
     def build_extension(self, ext: Extension):
         # Taken from setuptools:
         # https://github.com/pypa/setuptools/blob/245d72a8aa4d47e1811425213aba2a06a0bb64fa/setuptools/command/build_ext.py#L240-L241
-        ext._convert_pyx_sources_to_lang()
+        ext._convert_pyx_sources_to_lang()  # type: ignore[attr-defined]
 
         # Taken from distutils:
         # https://github.com/pypa/distutils/blob/4435cec31b8eb5712aa8bf993bea3f07051c24d8/distutils/command/build_ext.py#L504-L513
         sources = ext.sources
         if sources is None or not isinstance(sources, (list, tuple)):
-            raise DistutilsSetupError(
+            raise distutils.errors.DistutilsSetupError(
                 "in 'ext_modules' option (extension '%s'), "
                 "'sources' must be present and must be "
                 "a list of source filenames" % ext.name
@@ -56,8 +59,8 @@ class cabal_build_ext(build_ext):
 
         # Taken from setuptools:
         # https://github.com/pypa/setuptools/blob/245d72a8aa4d47e1811425213aba2a06a0bb64fa/setuptools/command/build_ext.py#L247-L249
-        if ext._needs_stub:
-            build_lib = self.get_finalized_command("build_py").build_lib
+        if ext._needs_stub:  # type: ignore[attr-defined]
+            build_lib = self.get_finalized_command("build_py").build_lib  # type: ignore[attr-defined]
             self.write_stub(build_lib, ext)
 
     def cabal_configure_ext(self):
@@ -80,7 +83,9 @@ class cabal_build_ext(build_ext):
 
     def get_cabal_foreign_library_filename(self, ext):
         if sys.platform not in ["darwin", "linux", "win32", "cygwin"]:
-            raise DistutilsPlatformError(f"unsupported platform {self.plat_name}")
+            raise distutils.errors.DistutilsPlatformError(
+                f"unsupported platform {self.plat_name}"
+            )
         library_prefix = "" if sys.platform in ["win32", "cygwin"] else "lib"
         component_name = ext.name.split(".")[-1]
         dynlib_extension = {
@@ -97,15 +102,19 @@ class cabal_build_ext(build_ext):
         print(cmd)
         exitCode = subprocess.call(args, env=env)
         if exitCode != 0:
-            raise DistutilsExecError(f"error occurred when running '{cmd}'")
+            raise distutils.errors.DistutilsExecError(
+                f"error occurred when running '{cmd}'"
+            )
 
     _cabal: Optional[str] = None
 
     def find_cabal(self):
         if self._cabal is None:
-            self._cabal = find_executable("cabal")
+            self._cabal = distutils.spawn.find_executable("cabal")
             if self._cabal is None:
-                raise DistutilsExecError("could not find executable 'cabal'")
+                raise distutils.errors.DistutilsExecError(
+                    "could not find executable 'cabal'"
+                )
         return self._cabal
