--- vehicle-python-core/src/vehicle/core/__main__.py
@@ -1,5 +1,0 @@
-import sys
-import vehicle.core
-
-if __name__ == "__main__":
-    vehicle.core.main(sys.argv)
