+++ vehicle-python-core/src/vehicle_core/__init__.py
@@ -1,0 +1,11 @@
+from typing import List
+
+from vehicle_core._binding import hs_defaultMain as _hs_defaultMain
+from vehicle_core._binding import hs_rts_exit as _hs_rts_exit
+from vehicle_core._binding import hs_rts_init as _hs_rts_init
+
+
+def main(args: List[str]):
+    _hs_rts_init(["vehicle"])
+    _hs_defaultMain(args)
+    _hs_rts_exit()
