+++ vehicle-python-core/src/vehicle_core/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+import vehicle_core
+
+if __name__ == "__main__":
+    vehicle_core.main(sys.argv)
