+++ vehicle-python-core/src/vehicle_core/_binding.pyi
