+++ vehicle/src/Vehicle/Compile/ExpandResources/Dataset.hs
@@ -4,6 +4,7 @@ module Vehicle.Compile.ExpandResources.Dataset
 where
 
 import Control.Monad.Except (MonadError (..))
+import Control.Monad.IO.Class (MonadIO)
 import Data.Map qualified as Map
 import System.FilePath (takeExtension)
 import Vehicle.Compile.Error
@@ -11,7 +12,6 @@ import Vehicle.Compile.ExpandResources.Core
 import Vehicle.Compile.ExpandResources.Dataset.IDX (readIDX)
 import Vehicle.Compile.Prelude
 import Vehicle.Compile.Type.Subsystem.Standard
-import Control.Monad.IO.Class (MonadIO)
 
 --------------------------------------------------------------------------------
 -- Dataset parsing
